requirements.txt
certifi==2023.7.22
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.1
filelock==3.11.0
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.3
lit==16.0.0
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
Pillow==10.2.0
requests==2.31.0
segment-anything==1.0
sympy==1.11.1
torch==2.0.0
torchvision==0.15.1
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.18
Werkzeug==2.3.8